ThumbnailImageCreator service introduced
