+++ src/Imaging/Extraction/ThumbnailImageCreatorInterface.php
@@ -1,0 +1,21 @@
+<?php
+/*
+ * This file is part of ImgCache.
+ *
+ * (c) Igor Lazarev <strider2038@rambler.ru>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Strider2038\ImgCache\Imaging\Extraction;
+
+use Strider2038\ImgCache\Imaging\Image\Image;
+
+/**
+ * @author Igor Lazarev <strider2038@rambler.ru>
+ */
+interface ThumbnailImageCreatorInterface
+{
+    public function createThumbnailImageByConfiguration(Image $image, string $processingConfiguration): Image;
+}
