+++ tests/Unit/Imaging/Extraction/ThumbnailImageExtractorTest.php
@@ -11,6 +11,7 @@
 namespace Strider2038\ImgCache\Tests\Unit\Imaging\Extraction;
 
 use PHPUnit\Framework\TestCase;
+use Strider2038\ImgCache\Imaging\Extraction\ThumbnailImageCreatorInterface;
 use Strider2038\ImgCache\Imaging\Extraction\ThumbnailImageExtractor;
 use Strider2038\ImgCache\Imaging\Image\Image;
 use Strider2038\ImgCache\Imaging\Image\ImageParameters;
@@ -31,21 +32,17 @@ class ThumbnailImageExtractorTest extends TestCase
     /** @var ThumbnailFilenameParserInterface */
     private $filenameParser;
 
-    /** @var ProcessingConfigurationParserInterface */
-    private $processingConfigurationParser;
-
     /** @var StorageAccessorInterface */
     private $storageAccessor;
 
-    /** @var ImageProcessorInterface */
-    private $imageProcessor;
+    /** @var ThumbnailImageCreatorInterface */
+    private $thumbnailImageCreator;
 
     protected function setUp(): void
     {
         $this->filenameParser = \Phake::mock(ThumbnailFilenameParserInterface::class);
-        $this->processingConfigurationParser = \Phake::mock(ProcessingConfigurationParserInterface::class);
         $this->storageAccessor = \Phake::mock(StorageAccessorInterface::class);
-        $this->imageProcessor = \Phake::mock(ImageProcessorInterface::class);
+        $this->thumbnailImageCreator = \Phake::mock(ThumbnailImageCreatorInterface::class);
     }
 
     /** @test */
@@ -54,10 +51,7 @@ class ThumbnailImageExtractorTest extends TestCase
         $extractor = $this->createThumbnailImageExtractor();
         $thumbnailFilename = $this->givenFilenameParser_getParsedFilename_returnsThumbnailFilename();
         $sourceImage = $this->givenStorageAccessor_getImage_returnsImage();
-        $processingConfiguration = $this->givenProcessingConfigurationParser_parseConfiguration_returnsProcessingConfiguration();
-        $transformations = $this->givenProcessingConfiguration_getTransformations_returnsTransformationCollection($processingConfiguration);
-        $imageParameters = $this->givenProcessingConfiguration_getImageParameters_returnsImageParameters($processingConfiguration);
-        $processedImage = $this->givenImageProcessor_transformImage_returnsImage();
+        $thumbnailImage = $this->givenThumbnailImageCreator_createThumbnailImageByConfiguration_returnsImage();
 
         $extractedImage = $extractor->getProcessedImage(self::FILENAME);
 
@@ -65,21 +59,19 @@ class ThumbnailImageExtractorTest extends TestCase
         $this->assertFilenameParser_getParsedFilename_isCalledOnceWithFilename(self::FILENAME);
         $this->assertStorageAccessor_getImage_isCalledOnceWithFilename(self::THUMBNAIL_FILENAME_VALUE);
         $this->assertThumbnailFilename_getProcessingConfiguration_isCalledOnce($thumbnailFilename);
-        $this->assertProcessingConfigurationParser_parseConfiguration_isCalledOnceWithConfiguration(self::PROCESSING_CONFIGURATION);
-        $this->assertProcessingConfiguration_getTransformations_isCalledOnce($processingConfiguration);
-        $this->assertProcessingConfiguration_getImageParameters_isCalledOnce($processingConfiguration);
-        $this->assertImageProcessor_transformImage_isCalledOnceWithImageAndTransformationCollection($sourceImage, $transformations);
-        $this->assertImage_setParameters_isCalledOnceWithParameters($processedImage, $imageParameters);
-        $this->assertSame($processedImage, $extractedImage);
+        $this->assertThumbnailImageCreator_createThumbnailImageByConfiguration_isCalledOnceWithImageAndConfiguration(
+            $sourceImage,
+            self::PROCESSING_CONFIGURATION
+        );
+        $this->assertSame($thumbnailImage, $extractedImage);
     }
 
     private function createThumbnailImageExtractor(): ThumbnailImageExtractor
     {
         return new ThumbnailImageExtractor(
             $this->filenameParser,
-            $this->processingConfigurationParser,
             $this->storageAccessor,
-            $this->imageProcessor
+            $this->thumbnailImageCreator
         );
     }
 
@@ -101,25 +93,15 @@ class ThumbnailImageExtractorTest extends TestCase
         return $thumbnailFilename;
     }
 
-    private function givenImageProcessor_transformImage_returnsImage(): Image
+    private function givenThumbnailImageCreator_createThumbnailImageByConfiguration_returnsImage(): Image
     {
         $image = \Phake::mock(Image::class);
 
-        \Phake::when($this->imageProcessor)
-            ->transformImage(\Phake::anyParameters())
+        \Phake::when($this->thumbnailImageCreator)
+            ->createThumbnailImageByConfiguration(\Phake::anyParameters())
             ->thenReturn($image);
 
         return $image;
-    }
-
-    private function givenProcessingConfigurationParser_parseConfiguration_returnsProcessingConfiguration(): ProcessingConfiguration
-    {
-        $processingConfiguration = \Phake::mock(ProcessingConfiguration::class);
-        \Phake::when($this->processingConfigurationParser)
-            ->parseConfiguration(\Phake::anyParameters())
-            ->thenReturn($processingConfiguration);
-
-        return $processingConfiguration;
     }
 
     private function assertFilenameParser_getParsedFilename_isCalledOnceWithFilename(string $filename): void
@@ -132,11 +114,12 @@ class ThumbnailImageExtractorTest extends TestCase
         \Phake::verify($this->storageAccessor, \Phake::times(1))->getImage($filename);
     }
 
-    private function assertImageProcessor_transformImage_isCalledOnceWithImageAndTransformationCollection(
+    private function assertThumbnailImageCreator_createThumbnailImageByConfiguration_isCalledOnceWithImageAndConfiguration(
         Image $image,
-        TransformationCollection $transformationCollection
+        string $configuration
     ): void {
-        \Phake::verify($this->imageProcessor, \Phake::times(1))->transformImage($image, $transformationCollection);
+        \Phake::verify($this->thumbnailImageCreator, \Phake::times(1))
+            ->createThumbnailImageByConfiguration($image, $configuration);
     }
 
     private function assertThumbnailFilename_getValue_isCalledOnce(ThumbnailFilename $thumbnailFilename): void
@@ -147,48 +130,5 @@ class ThumbnailImageExtractorTest extends TestCase
     private function assertThumbnailFilename_getProcessingConfiguration_isCalledOnce(ThumbnailFilename $thumbnailFilename): void
     {
         \Phake::verify($thumbnailFilename, \Phake::times(1))->getProcessingConfiguration();
-    }
-
-    private function assertProcessingConfigurationParser_parseConfiguration_isCalledOnceWithConfiguration(
-        string $configuration
-    ): void {
-        \Phake::verify($this->processingConfigurationParser, \Phake::times(1))->parseConfiguration($configuration);
-    }
-
-    private function assertProcessingConfiguration_getTransformations_isCalledOnce(
-        ProcessingConfiguration $processingConfiguration
-    ): void {
-        \Phake::verify($processingConfiguration, \Phake::times(1))->getTransformations();
-    }
-
-    private function assertProcessingConfiguration_getImageParameters_isCalledOnce(
-        ProcessingConfiguration $processingConfiguration
-    ): void {
-        \Phake::verify($processingConfiguration, \Phake::times(1))->getImageParameters();
-    }
-
-    private function givenProcessingConfiguration_getTransformations_returnsTransformationCollection(
-        ProcessingConfiguration $processingConfiguration
-    ): TransformationCollection {
-        $transformations = \Phake::mock(TransformationCollection::class);
-        \Phake::when($processingConfiguration)->getTransformations()->thenReturn($transformations);
-
-        return $transformations;
-    }
-
-    private function givenProcessingConfiguration_getImageParameters_returnsImageParameters(
-        ProcessingConfiguration $processingConfiguration
-    ): ImageParameters {
-        $parameters = \Phake::mock(ImageParameters::class);
-        \Phake::when($processingConfiguration)->getImageParameters()->thenReturn($parameters);
-
-        return $parameters;
-    }
-
-    private function assertImage_setParameters_isCalledOnceWithParameters(
-        Image $image,
-        ImageParameters $imageParameters
-    ): void {
-        \Phake::verify($image, \Phake::times(1))->setParameters($imageParameters);
     }
 }
